requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

msedge_selenium_tools == 3.141.2
pytest == 6.1.1
selenium == 3.141.0
